Implement QR Code Builder for HOTP & Reorganize Codebase
- [+] refactor: move blake2b hash functions to separate package blake2botp

- [+] feat(otpverifier): add HOTP verifier implementation
- [+] Implement HOTPVerifier struct
- [+] Add NewHOTPVerifier constructor
- [+] Implement Verify and GenerateToken methods for HOTP
- [+] Add generateSignature method for HOTP

- [+] feat(otpverifier): add QRCodeConfig struct and default configuration
- [+] Define QRCodeConfig struct for QR code generation configuration
- [+] Add DefaultQRCodeConfig variable with default configuration values
- [+] Implement InitializeDefaultQRCodeConfig function to set up default configuration with dynamic text positions

- [+] refactor(otpverifier): move QR code generation functions to separate file
- [+] Move BuildQRCode and SaveQRCodeImage methods for TOTPVerifier to qrcode_builder.go
- [+] Move ensureDefaultConfig and generateOTPURL functions to qrcode_builder.go

- [+] feat(otpverifier): add QR code generation support for HOTPVerifier
- [+] Implement BuildQRCode method for HOTPVerifier
- [+] Implement SaveQRCodeImage method for HOTPVerifier
- [+] Update generateOTPURL function to handle HOTP configuration

- [+] test(otpverifier): add tests for HOTP verifier and QR code generation
- [+] Add TestHOTPVerifier_BuildQRCode test
- [+] Add TestHOTPVerifier_SaveQRCodeImage test

- [+] refactor: move TOTP verifier implementation to separate file totp.go
